backend/requirements.txt
# FastAPI 및 ASGI 서버
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 데이터베이스
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# 인증
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AWS SDK (Nova-2 STT용)
boto3==1.34.34

# LLM API
openai==1.10.0
anthropic==0.8.1
langsmith==0.0.87
langchain>=0.1.0
langchain-community>=0.0.1
langchain-core>=0.1.0

# WebSocket (실시간 통신)
websockets==12.0

# 데이터 처리
pandas==2.2.0
openpyxl==3.1.2

# 환경 변수 관리
python-dotenv==1.0.1

# 데이터 검증
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP 클라이언트
httpx==0.26.0
aiofiles==23.2.1

# 테스트
pytest==7.4.4
pytest-asyncio==0.23.3

# 유틸리티
python-dateutil==2.8.2
